requirements.txt
boto==2.42.0
boto3==1.4.0
botocore==1.4.54
Django==1.10.1
django-storages==1.5.1
docutils==0.12
jmespath==0.9.0
Pillow==3.3.1
psycopg2==2.6.2
PyPDF2==1.26.0
pytesseract==0.1.6
python-dateutil==2.5.3
s3transfer==0.1.4
six==1.10.0
Wand==0.4.3
